requirements.txt
# ==============================
# EMERGENCE - Requirements (E2E)
# Targets: Windows dev + Cloud Run
# ==============================

# --- Core server ---
fastapi==0.119.0                 # upgraded for async improvements + dependency cleanup fixes
uvicorn[standard]==0.30.1        # uvloop/httptools included for perf
websockets==15.0.1               # pinned for reproducibility
python-multipart==0.0.20         # pinned for reproducibility
PyJWT>=2.9,<3
# --- Auth ---
bcrypt>=4.1,<5
pyotp>=2.9,<3  # TOTP 2FA (Phase P2 - Feature 9)
qrcode>=8.0,<9  # QR code generation for 2FA setup

# --- DI & Config ---
dependency-injector>=4.41,<5
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
python-dotenv>=1.0,<2

# --- LLM providers ---
openai>=1.35,<2
google-generativeai==0.8.5       # match env, avoid breaking changes
anthropic==0.64.0                # match env stable
tiktoken>=0.5.2,<0.8

# --- Data & vectors ---
google-cloud-firestore>=2.16,<3
google-cloud-storage>=2.10,<3  # Guardian reports Cloud Storage
google-cloud-logging>=3.5,<4   # Guardian production logs analysis
aiosqlite==0.21.0                # match env stable
chromadb==0.5.23                 # upgraded for batch upsert optimizations (+20-40% performance)
qdrant-client>=1.7.3,<2          # optional vector backend (Qdrant)

# --- Embeddings & NumPy/Torch ---
sentence-transformers>=2.7,<3
# Torch CPU by default; override if CUDA is required in your environment
torch>=2.1,<3
numpy==1.26.4                    # matches env; <2 for torch compatibility

# --- Files & utilities ---
aiofiles==24.1.0                 # match env stable
PyMuPDF>=1.24.10,<1.25           # recent fitz/PyMuPDF
python-docx>=1.1.0,<1.2

# --- Email templates ---
Jinja2>=3.1,<4                   # Email templates rendering

# --- Gmail API (Phase 3 Guardian Cloud) ---
google-auth>=2.23,<3
google-auth-oauthlib>=1.1,<2
google-auth-httplib2>=0.2,<1
google-api-python-client>=2.108,<3
google-cloud-secret-manager>=2.16,<3  # OAuth client secret storage

# OS specific (magic/libmagic)
python-magic>=0.4.27; platform_system!="Windows"
python-magic-bin>=0.4.14; platform_system=="Windows"

chardet==5.2.0
psutil>=5.9,<6

# --- Monitoring ---
prometheus-client>=0.20,<1
httpx>=0.24,<0.28                # Required by VoiceService (containers.py) / TestClient compat

# --- Tooling & QA ---
pytest>=8.3,<9                   # ensure unit tests run in CI/local
pytest-asyncio>=1.2,<2           # async test support
ruff>=0.13,<1                   # lint backend modules (Pydantic v2, style)
mypy>=1.18,<2                   # static typing for backend
types-psutil>=7.0.0             # stubs for psutil (mypy clean)

# --- Benchmarks & Analytics ---
PyYAML>=6.0,<7                  # memory_probe.py ground truth loading
matplotlib>=3.7,<4              # retention curve plotting
pandas>=2.0,<3                  # CSV aggregation & pivot tables
